API

In the following, you find the documentation of all exported functions of the MRFingerprintingRecon.jl package:

MRFingerprintingRecon.FFTNormalOp

MRFingerprintingRecon.NFFTNormalOp

MRFingerprintingRecon.calculateBackProjection

MRFingerprintingRecon.calculateBackProjection_gridded

MRFingerprintingRecon.calculateGoldenMeans

MRFingerprintingRecon.grog_calib

MRFingerprintingRecon.grog_gridding!

MRFingerprintingRecon.kooshball

MRFingerprintingRecon.kooshballGA

MRFingerprintingRecon.radial_grog!

MRFingerprintingRecon.traj_2d_cartesian

MRFingerprintingRecon.traj_2d_radial_goldenratio

MRFingerprintingRecon.FFTNormalOp — Method

FFTNormalOp(img_shape, trj, U; cmaps)
FFTNormalOp(M, U; cmaps)
FFTNormalOp(Λ; cmaps)

Create normal operator of FFT operator. Differentiate between functions exploiting a pre-calculated kernel basis Λ and the functions which calculate Λ based on a passed trajectory trj or mask M.

Arguments

img_shape::Tuple{Int}: Image dimensions

traj::Vector{Matrix{Float32}}: Trajectory

U::Matrix{ComplexF32}: Basis coefficients of subspace

cmaps::Matrix{ComplexF32}: Coil sensitivities

M::Vector{Matrix{Float32}}: Mask

Λ::Array{Complex{T},3}: Toeplitz kernel basis

MRFingerprintingRecon.NFFTNormalOp — Method

NFFTNormalOp(img_shape, trj, U; cmaps, verbose, num_fft_threads)
NFFTNormalOp(img_shape, Λ, kmask_indcs; cmaps)

Create normal operator of NFFT operator. Differentiate between functions exploiting a pre-calculated Toeplitz kernel basis Λ and the function which calculates Λ based on a passed trajectory trj.

Arguments

img_shape::Tuple{Int}: Image dimensions

traj::Vector{Matrix{Float32}}: Trajectory

U::Matrix{ComplexF32}: Basis coefficients of subspace

cmaps::Matrix{ComplexF32}: Coil sensitivities

Λ::Array{Complex{T},3}: Toeplitz kernel basis

kmask_indcs::Vector{Int}: Sampling indices of Toeplitz mask

verbose::Boolean: Verbose level

num_fft_threads::Int: Number of threads for FFT

MRFingerprintingRecon.calculateBackProjection — Method

calculateBackProjection(data, trj, img_shape; U, density_compensation, verbose)
calculateBackProjection(data, trj, cmaps; U, density_compensation, verbose)

Calculate backprojection

Arguments

data::AbstractArray{T}: Basis coefficients of subspace

trj::Vector{Matrix{Float32}}: Trajectory

img_shape::NTuple{N,Int}: Shape of image

U::Matrix{ComplexF32}: Basis coefficients of subspace

density_compensation: Values of :radial_3D, :radial_2D, :none, or of type AbstractVector{<:AbstractVector}`

verbose::Boolean: Verbosity level

cmaps::::AbstractVector{<:AbstractArray{T}}: Coil sensitivities

MRFingerprintingRecon.calculateBackProjection_gridded — Method

calculateBackProjection_gridded(data, trj, U, cmaps)

Calculate gridded backprojection

Arguments

data::Matrix{ComplexF32}: Basis coefficients of subspace

trj::Vector{Matrix{Float32}}: Trajectory

U::Matrix{ComplexF32}: Basis coefficients of subspace

cmaps::Matrix{ComplexF32}: Coil sensitivities

Note

In case of repeated sampling (Nrep > 1), a joint basis reconstruction is required. Therefore, the basis needs to have a temporal dimension of Nt⋅Nrep with Nt as time dimension defined by the trajectory.

MRFingerprintingRecon.calculateGoldenMeans — Method

calculateGoldenMeans()

Function to calculate 3D golden means [1].

References

[1] Chan, R.W., Ramsay, E.A., Cunningham, C.H. and Plewes, D.B. (2009), Temporal stability of adaptive 3D radial MRI using multidimensional golden means. Magn. Reson. Med., 61: 354-363. https://doi.org/10.1002/mrm.21837

MRFingerprintingRecon.grog_calib — Method

grog_calib(data, trj, Nr)

Perform GROG kernel calibration based on whole radial trajectory and passed data. Calibration follows the work on self-calibrating radial GROG (https://doi.org/10.1002/mrm.21565).

Arguments

data::Matrix{ComplexF32}: Basis coefficients of subspace

traj::Vector{Matrix{Float32}}: Trajectory

Nr::Int: Number of samples per read out

MRFingerprintingRecon.grog_gridding! — Method

grog_gridding!(data, trj, lnG, Nr, img_shape)

Perform gridding of data based on pre-calculated GROG kernel.

Arguments

data::Matrix{ComplexF32}: Basis coefficients of subspace

traj::Vector{Matrix{Float32}}: Trajectory

lnG::Vector{Matrix{Float32}}: Natural logarithm of GROG kernel in all dimensions

Nr::Int: Number of samples per read out

img_shape::Tuple{Int}: Image dimensions

Dimensions:

data: [samples, spokes, timesteps, coils, repetitions of sampling pattern]

trj: [timesteps, repetitions][dims, samples]

lnG: [dims][Ncoils, Ncoils]

Further:

Ensure sampling pattern repeats in repetitions dimension!

MRFingerprintingRecon.kooshball — Method

kooshball(Nr, theta, phi; thetaRot = 0, phiRot = 0, delay = (0, 0, 0), T = Float32)

Function to calculate kooshball trajectory.

Arguments

Nr::Int: Number of read out samples

theta::Array{Float,2}: Array with dimensions: Ncyc, Nt defining the angles theta for each cycle and timestep.

phi::Array{Float,2}: Array with dimensions: Ncyc, Nt defining the angles phi for each cycle and timestep.

thetaRot::Float: Fixed rotation angle along theta

phiRot::Float: Fixed rotation angle along phi

delay::Tuple{Float, Float, Float}: Gradient delays in (HF, AP, LR)

T::Type: Type defining the output units of the trajectory

MRFingerprintingRecon.kooshballGA — Method

kooshballGA(Nr, Ncyc, Nt; thetaRot = 0, phiRot = 0, delay = (0, 0, 0), T = Float32)

Function to calculate golden means [1] based kooshball trajectory.

Arguments

Nr::Int: Number of read out samples

Ncyc::Int: Number of cycles

Nt::Int: Number of time steps in the trajectory

thetaRot::Float: Fixed rotation angle along theta

phiRot::Float: Fixed rotation angle along phi

delay::Tuple{Float, Float, Float}: Gradient delays in (HF, AP, LR)

T::Type: Type defining the output units of the trajectory

References

[1] Chan, R.W., Ramsay, E.A., Cunningham, C.H. and Plewes, D.B. (2009), Temporal stability of adaptive 3D radial MRI using multidimensional golden means. Magn. Reson. Med., 61: 354-363. https://doi.org/10.1002/mrm.21837

MRFingerprintingRecon.radial_grog! — Method

radial_grog!(data, trj, Nr, img_shape)

Perform GROG kernel calibration and gridding of data in-place.

Arguments

data::Matrix{ComplexF32}: Basis coefficients of subspace

traj::Vector{Matrix{Float32}}: Trajectory

Nr::Int: Number of samples per read out

img_shape::Tuple{Int}: Image dimensions

MRFingerprintingRecon.traj_2d_cartesian — Method

traj_2d_cartesian(Nx, Ny, Nz, Nt; samplingRate_units = true, T = Float32)

Function to calculate a 2D cartesian trajectory in units of sampling rate ∈ {x | -N/2+1 ≤ x ≤ N/2 and x ∈ Z}. With samplingRate_units = false the ouput is relative with samples ∈ [-1/2:1/2].

Arguments

Nx::Int: Number of frequency encoded samples per read out

Ny::Int: Number of phase encoding lines

Nz::Int: Number of phase encoding lines (third dimension)

Nt::Int: Number of times the sampling pattern is repeated

samplingRate_units::Boolean: Parameter setting the output units to sampling rate

T::Type: Type defining the output units of the trajectory

MRFingerprintingRecon.traj_2d_radial_goldenratio — Method

traj_2d_radial_goldenratio(Nr, Ncyc, Nt; thetaRot = 0, phiRot = 0, delay = (0, 0, 0), N = 1, T = Float32)

Function to calculate 2D golden ratio based trajectory [1]. By modifying N also tiny golden angles [2] are supported.

Arguments

Nr::Int: Number of read out samples

Ncyc::Int: Number of cycles

Nt::Int: Number of time steps in the trajectory

thetaRot::Float: Fixed rotation angle along theta

phiRot::Float: Fixed rotation angle along phi

delay::Tuple{Float, Float, Float}: Gradient delays in (HF, AP, LR)

N::Int: Number of tiny golden angle

T::Type: Type defining the output units of the trajectory

References

[1] Winkelmann S, Schaeffter T, Koehler T, Eggers H, Doessel O. An optimal radial profile order based on the Golden Ratio for time-resolved MRI. IEEE TMI 26:68–76 (2007) [2] Wundrak S, Paul J, Ulrici J, Hell E, Geibel MA, Bernhardt P, Rottbauer W, Rasche V. Golden ratio sparse MRI using tiny golden angles. Magn Reson Med 75:2372-2378 (2016)